backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to version compatible with passlib
slowapi==0.1.9

# Request Handling
python-multipart==0.0.20
pydantic-settings==2.1.0
sse-starlette==1.8.2

# File System & Scheduling
watchdog==3.0.0
apscheduler==3.10.4

# Browser Automation
playwright==1.40.0

# Audio/Video Processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1

# Image Processing
Pillow==10.1.0

# YouTube Integration
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# HTTP & File I/O
python-dotenv==1.0.0
requests==2.32.5
aiohttp==3.13.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
